+++ src/core/net/ip6.cpp
@@ -77,7 +77,7 @@ uint16_t Ip6::UpdateChecksum(uint16_t checksum, const void *buf, uint16_t len)
 
     for (int i = 0; i < len; i++)
     {
-        checksum = Ip6::UpdateChecksum(checksum, (i & 1) ? bytes[i] : (static_cast<uint16_t>(bytes[i])) << 8);
+        checksum = Ip6::UpdateChecksum(checksum, (i & 1) ? bytes[i] : static_cast<uint16_t>(bytes[i] << 8));
     }
 
     return checksum;
@@ -335,7 +335,7 @@ exit:
     }
 }
 
-ThreadError Ip6::HandleDatagram(Message &message, Netif *netif, uint8_t interfaceId, const void *linkMessageInfo,
+ThreadError Ip6::HandleDatagram(Message &message, Netif *netif, int8_t interfaceId, const void *linkMessageInfo,
                                 bool fromLocalHost)
 {
     ThreadError error = kThreadError_Drop;
@@ -455,7 +455,7 @@ exit:
 ThreadError ForwardMessage(Message &message, MessageInfo &messageInfo)
 {
     ThreadError error = kThreadError_None;
-    int interfaceId;
+    int8_t interfaceId;
     Netif *netif;
 
     if (messageInfo.GetSockAddr().IsMulticast())
